1x 599 rating script updated

+++ mggk_bash_scripts/599_mggk_REPLACE_BULK_RATINGCOUNT_AND_RATINGVALUE_FROM_GOOGLE_SHEETS_CSV.sh
@@ -43,21 +43,20 @@ if [ "$1" == "--help" ] ; then usage ; fi
 ##++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 ################################################################################
-#HUGO_CONTENT_DIR="$HOME/GitHub/2019-HUGO-MGGK-WEBSITE-OFFICIAL/content/_FIXED" ;
-HUGO_CONTENT_DIR="$HOME/GitHub/2019-HUGO-MGGK-WEBSITE-OFFICIAL/content/popular-posts" ;
+HUGO_CONTENT_DIR="$HOME/GitHub/2019-HUGO-MGGK-WEBSITE-OFFICIAL/content" ;
 MY_PWD="$HOME_WINDOWS/Desktop/Y"
 cd $MY_PWD ;
 echo "Current working directory = $MY_PWD" ;
 ################################################################################
 ## SETTING VARIABLES
-TMP_OUTPUT_CSVFILE="$MY_PWD/_TMP_OUTPUT_598_MGGK_RATING_VALID_URLS.csv"
+TMP_OUTPUT_CSVFILE="$MY_PWD/_TMP_OUTPUT_SUMMARY_599_MGGK_RATING_VALID_URLS.csv"
 ## INITIALIZING TMP OUTPUT CSV FILE
-echo "MY_MDFILENAME, COUNT, CSVURL, CSVRATINGCOUNT, CSVRATINGVALUE" > $TMP_OUTPUT_CSVFILE
+echo "MY_MDFILENAME, COUNT, CSVURL, 90DAYS_PAGEVIEWS, EXISTING_CSVRATINGCOUNT, NEW_RATINGCOUNT_INCREASED_BY, NEW_RATINGCOUNT, EXISTING_CSVRATINGVALUE, NEW_RATINGVALUE" > $TMP_OUTPUT_CSVFILE
 ################################################################################
 ## CALCULATION VARIABLES (CHANGE AS NEEDED)
 NUM_DAYS_DATA_IN_REQUIREMENTS_CSVFILE="90" ;
 PERCENTAGE_OF_USERS_WHO_RATED="0.04" ; ## Let's say that 4% users provided rating out of total users
-NEW_RATINGVALUE_CHANGE_FACTOR="4.6" ; ## Let's say that average rating by new users is 4.5
+NEW_RATINGVALUE_CHANGE_FACTOR="4.7" ; ## Let's say that average rating by new users is 4.7
 ################################################################################
 
 ## GETTING COLUMN NAMES FROM CSV FILE:
@@ -106,23 +105,35 @@ REPLACE_ORIGINAL_RATINGCOUNT_RATINGVALUE_IN_THIS_MD_FILE_WITH_NEW_VALUES () {
   mdfile="$1" ;
   EXISTING_RATINGCOUNT="$(grep -irh 'ratingCount: ' $mdfile | sed 's/ratingCount: //g')" ;
   NEW_RATINGCOUNT_TMP="$2" ;
-  NEW_RATINGCOUNT_INCREASED_BY=$(echo "(($NEW_RATINGCOUNT_TMP * $DAYS_SINCE_WEBSITE_LAST_UPDATED * $PERCENTAGE_OF_USERS_WHO_RATED) / $NUM_DAYS_DATA_IN_REQUIREMENTS_CSVFILE)"  | bc ) ;
+
+  CALCULATING_RATINGCOUNT_INCREASED_BY="(($NEW_RATINGCOUNT_TMP * $DAYS_SINCE_WEBSITE_LAST_UPDATED * $PERCENTAGE_OF_USERS_WHO_RATED) / $NUM_DAYS_DATA_IN_REQUIREMENTS_CSVFILE)" ;
+  echo "CALCULATING_RATINGCOUNT_INCREASED_BY = $CALCULATING_RATINGCOUNT_INCREASED_BY" ;
+  NEW_RATINGCOUNT_INCREASED_BY=$(echo "$CALCULATING_RATINGCOUNT_INCREASED_BY"  | bc ) ;
   NEW_RATINGCOUNT=$(echo "$EXISTING_RATINGCOUNT + $NEW_RATINGCOUNT_INCREASED_BY"  | bc ) ;
 
   echo "    mdfile => $mdfile" ;
   echo "    EXISTING_RATINGCOUNT => $EXISTING_RATINGCOUNT" ;
   echo "    NEW_RATINGCOUNT_INCREASED_BY => $NEW_RATINGCOUNT_INCREASED_BY" ;
   echo "    NEW_RATINGCOUNT => $NEW_RATINGCOUNT" ;
-  echo "    >>>> Updating new ratingCount value ..." ;
-  sed -i .bak "s|ratingCount: .*$|ratingCount: $NEW_RATINGCOUNT|" $mdfile ;
-  rm $mdfile.bak ;
-  echo;
   ##
+  echo "    >>>> Updating new ratingCount value ..." ;
+  ## Check the OS // MacOS (=Darwin) as well as Linux (=ubuntu)
+  OS_NAME=$(uname) ;
+  if [[ "$OS_NAME" == "Darwin" ]] ; then
+    sed -i .bak "s|ratingCount: .*$|ratingCount: $NEW_RATINGCOUNT|" $mdfile ;
+    rm $mdfile.bak ;
+  else
+    sed -i "s|ratingCount: .*$|ratingCount: $NEW_RATINGCOUNT|" $mdfile ;
+  fi
+  ##
+  echo;
   ################## RATING VALUE CALCULATIONS #######################
   EXISTING_RATINGVALUE="$(grep -irh 'ratingValue: ' $mdfile | sed 's/ratingValue: //g')" ;
   ## Only change ratingValue when there is a change in ratingCount
   if [ $NEW_RATINGCOUNT_INCREASED_BY != 0 ]; then
-    NEW_RATINGVALUE=$(echo "scale=1; ($EXISTING_RATINGVALUE + $NEW_RATINGVALUE_CHANGE_FACTOR) / 2"  | bc ) ;
+    CALCULATING_RATINGVALUE="(($EXISTING_RATINGVALUE * $EXISTING_RATINGCOUNT) + ( $NEW_RATINGVALUE_CHANGE_FACTOR * $NEW_RATINGCOUNT_INCREASED_BY )) / ( $EXISTING_RATINGCOUNT + $NEW_RATINGCOUNT_INCREASED_BY)" ; 
+    echo "CALCULATING_RATINGVALUE = $CALCULATING_RATINGVALUE" ;
+    NEW_RATINGVALUE=$(echo "scale=1; $CALCULATING_RATINGVALUE"  | bc ) ;
   else 
     NEW_RATINGVALUE="$EXISTING_RATINGVALUE" ;
   fi 
@@ -130,10 +141,21 @@ REPLACE_ORIGINAL_RATINGCOUNT_RATINGVALUE_IN_THIS_MD_FILE_WITH_NEW_VALUES () {
   echo "    mdfile => $mdfile" ;
   echo "    EXISTING_RATINGVALUE => $EXISTING_RATINGVALUE" ;
   echo "    NEW_RATINGVALUE => $NEW_RATINGVALUE" ;
-  #echo "    >>>> Updating new ratingValue value ..." ;
-  #sed -i .bak "s|ratingValue: .*$|ratingValue: $NEW_RATINGVALUE|" $mdfile ;
-  #rm $mdfile.bak ;
+  ##
+  echo "    >>>> Updating new ratingValue value ..." ;
+  ## Check the OS // MacOS (=Darwin) as well as Linux (=ubuntu)
+  OS_NAME=$(uname) ;
+  if [[ "$OS_NAME" == "Darwin" ]] ; then
+    sed -i .bak "s|ratingValue: .*$|ratingValue: $NEW_RATINGVALUE|" $mdfile ;
+    rm $mdfile.bak ;
+  else
+    sed -i "s|ratingValue: .*$|ratingValue: $NEW_RATINGVALUE|" $mdfile ;
+  fi
+  ##
   echo;
+  ##
+  CSVURL="$3" ;
+  echo "\"$mdfile\",\"$COUNT\", \"$CSVURL\", \"$NEW_RATINGCOUNT_TMP\",\"$EXISTING_RATINGCOUNT\", \"$NEW_RATINGCOUNT_INCREASED_BY\", \"$NEW_RATINGCOUNT\", \"$EXISTING_RATINGVALUE\", \"$NEW_RATINGVALUE\"" >> $TMP_OUTPUT_CSVFILE
 }
 
 #### END: DEFINING MAIN FUNCTIONS ############################################
@@ -141,10 +163,16 @@ REPLACE_ORIGINAL_RATINGCOUNT_RATINGVALUE_IN_THIS_MD_FILE_WITH_NEW_VALUES () {
 
 ##################################################################################
 ## USER INPUT FOR THE TIME WHEN LAST TIME SITE WAS MADE
+echo "##################################################################################" ;
+say "Updating Rating Count and Rating Value ... How many days ago did you make the website last time (ENTER an integer) ... " ;
 echo ">> Now updating Rating Count and Rating Value ..." ;
 echo ">>>> How many days ago did you make the website last time (ENTER an integer) [0=today, 1=yesterday, 2=2 days ago, etc.]:" ;
-say "How many days ago did you make the website last time (ENTER an integer) ... " ;
 read DAYS_SINCE_WEBSITE_LAST_UPDATED
+## Exit this script if zero entered
+if [ $DAYS_SINCE_WEBSITE_LAST_UPDATED == 0 ] ; then 
+  echo ">>> NO CHANGE WILL BE DONE. This script will exit now."; 
+  exit 1 ; 
+fi
 ##################################################################################
 
 
@@ -173,8 +201,7 @@ while read line; do
     ########
     if [[ "$MY_MDFILENAME" != "" ]] ; then
         echo "  >>>> RUNNING FUNCTIONS ..." ;
-        REPLACE_ORIGINAL_RATINGCOUNT_RATINGVALUE_IN_THIS_MD_FILE_WITH_NEW_VALUES "$MY_MDFILENAME" "$CSVRATINGCOUNT"
-        echo "\"$MY_MDFILENAME\",\"$COUNT\", \"$CSVURL\", \"$CSVRATINGCOUNT\", \"$CSVRATINGVALUE\"" >> $TMP_OUTPUT_CSVFILE
+        REPLACE_ORIGINAL_RATINGCOUNT_RATINGVALUE_IN_THIS_MD_FILE_WITH_NEW_VALUES "$MY_MDFILENAME" "$CSVRATINGCOUNT" "$CSVURL"
     fi
     ########
